refactor: isolate gemini stream projection
