+++ crates/llm/src/google_gemini_generate_content.rs
@@ -4,7 +4,6 @@
 //! Authentication is handled via `gcloud auth print-access-token`.
 
 use anyhow::{Context, Result};
-use futures::StreamExt;
 use serde::{Deserialize, Serialize};
 use std::process::Command;
 use tracing::{debug, error, instrument, warn};
@@ -12,8 +11,10 @@ use tracing::{debug, error, instrument, warn};
 use crate::ReasoningEffort;
 
 mod input_projection;
+mod streaming;
 
 use input_projection::project_request_input;
+pub use streaming::{StreamCandidate, StreamChunk};
 
 /// Client for the Google Gemini GenerateContent API.
 pub struct GoogleGeminiGenerateContentClient {
@@ -350,52 +351,7 @@ impl GoogleGeminiGenerateContentClient {
             anyhow::bail!("Gemini streaming API error ({}): {}", status, error_text);
         }
 
-        // Process SSE stream with event-boundary parsing.
-        let mut stream = response.bytes_stream();
-        let mut parser = SseEventParser::new();
-
-        while let Some(chunk_result) = stream.next().await {
-            let chunk = chunk_result.context("Failed to read stream chunk")?;
-            for data in parser.push(&chunk) {
-                if data == "[DONE]" {
-                    debug!("Stream completed");
-                    return Ok(());
-                }
-
-                match serde_json::from_str::<StreamChunk>(&data) {
-                    Ok(stream_chunk) => {
-                        if tx.send(stream_chunk).await.is_err() {
-                            debug!("Receiver dropped, stopping stream");
-                            return Ok(());
-                        }
-                    }
-                    Err(e) => {
-                        debug!(?e, data, "Failed to parse stream chunk");
-                    }
-                }
-            }
-        }
-
-        for data in parser.finish() {
-            if data == "[DONE]" {
-                debug!("Stream completed");
-                return Ok(());
-            }
-
-            match serde_json::from_str::<StreamChunk>(&data) {
-                Ok(stream_chunk) => {
-                    if tx.send(stream_chunk).await.is_err() {
-                        debug!("Receiver dropped, stopping stream");
-                        return Ok(());
-                    }
-                }
-                Err(e) => {
-                    debug!(?e, data, "Failed to parse stream chunk");
-                }
-            }
-        }
-
-        Ok(())
+        streaming::consume_response_stream(response, tx).await
     }
 
     /// Simple chat helper - send message and get text response
@@ -463,35 +419,6 @@ impl GoogleGeminiGenerateContentClient {
 
         Ok(text)
     }
-}
-
-// ============================================================================
-// Streaming Response Types
-// ============================================================================
-
-/// Stream chunk from Gemini streaming API
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct StreamChunk {
-    /// Generated candidates
-    #[serde(default)]
-    pub candidates: Vec<StreamCandidate>,
-    /// Usage metadata (only present in final chunk)
-    pub usage_metadata: Option<UsageMetadata>,
-    /// Prompt feedback (e.g., blocked prompt) in some streamed responses.
-    pub prompt_feedback: Option<PromptFeedback>,
-}
-
-/// A candidate in streaming response
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct StreamCandidate {
-    /// Content of the response
-    pub content: Option<Content>,
-    /// Why generation stopped (only in final chunk)
-    pub finish_reason: Option<String>,
-    /// Index of this candidate
-    pub index: Option<i32>,
 }
 
 // ============================================================================
@@ -564,41 +491,9 @@ impl Content {
 // ============================================================================
 
 use crate::{
-    GenerationRequest, GenerationResponse, LlmProvider, SseEventParser,
-    StreamChunk as UnifiedStreamChunk, TokenUsage, ToolCall as LlmToolCall,
+    GenerationRequest, GenerationResponse, LlmProvider, StreamChunk as UnifiedStreamChunk,
+    TokenUsage, ToolCall as LlmToolCall,
 };
-
-fn select_stream_candidate_index(
-    selected_candidate_index: Option<i32>,
-    emitted_payload: bool,
-    candidates: &[StreamCandidate],
-) -> Option<i32> {
-    if candidates.is_empty() {
-        return selected_candidate_index;
-    }
-
-    let has_index_zero = candidates
-        .iter()
-        .any(|candidate| candidate.index == Some(0));
-    match selected_candidate_index {
-        Some(0) => Some(0),
-        Some(_current) if has_index_zero && !emitted_payload => Some(0),
-        Some(current) => Some(current),
-        None if has_index_zero => Some(0),
-        None => candidates.first().and_then(|candidate| candidate.index),
-    }
-}
-
-fn should_consume_stream_candidate(
-    selected_candidate_index: Option<i32>,
-    candidate_position: usize,
-    candidate_index: Option<i32>,
-) -> bool {
-    match selected_candidate_index {
-        Some(index) => candidate_index == Some(index),
-        None => candidate_position == 0,
-    }
-}
 
 fn select_primary_candidate(candidates: &[Candidate]) -> Option<&Candidate> {
     candidates
@@ -611,14 +506,6 @@ fn is_blocking_finish_reason(finish_reason: &str) -> bool {
     finish_reason.eq_ignore_ascii_case("SAFETY")
         || finish_reason.eq_ignore_ascii_case("RECITATION")
         || finish_reason.eq_ignore_ascii_case("OTHER")
-}
-
-fn normalize_stream_finish_reason(finish_reason: String) -> String {
-    if is_blocking_finish_reason(&finish_reason) {
-        format!("stream_error:{}", finish_reason.to_ascii_lowercase())
-    } else {
-        finish_reason
-    }
 }
 
 fn build_gemini_thinking_config(
@@ -861,164 +748,9 @@ impl LlmProvider for GoogleGeminiGenerateContentClient {
             }),
         };
 
-        let (gemini_tx, mut gemini_rx) = tokio::sync::mpsc::channel::<StreamChunk>(128);
+        let (gemini_tx, gemini_rx) = tokio::sync::mpsc::channel::<StreamChunk>(128);
         let (tx, rx) = tokio::sync::mpsc::channel(256);
-
-        // Convert Gemini chunks to unified chunks
-        let convert_tx = tx;
-        tokio::spawn(async move {
-            let mut latest_usage: Option<TokenUsage> = None;
-            let mut emitted_final = false;
-            let mut emitted_payload = false;
-            let mut selected_candidate_index: Option<i32> = None;
-            let mut next_tool_call_index: usize = 0;
-
-            while let Some(gemini_chunk) = gemini_rx.recv().await {
-                if let Some(usage) = gemini_chunk.usage_metadata {
-                    latest_usage = Some(TokenUsage {
-                        prompt_tokens: usage.prompt_token_count.unwrap_or(0),
-                        cached_prompt_tokens: None,
-                        completion_tokens: usage.candidates_token_count.unwrap_or(0),
-                        total_tokens: usage.total_token_count.unwrap_or(0),
-                        reasoning_tokens: None,
-                    });
-                }
-
-                if let Some(feedback) = gemini_chunk.prompt_feedback
-                    && let Some(block_reason) = feedback.block_reason
-                {
-                    let _ = convert_tx
-                        .send(UnifiedStreamChunk {
-                            text: None,
-                            thinking: None,
-                            thinking_signature: None,
-                            redacted_thinking: None,
-                            usage: latest_usage,
-                            sequence_number: None,
-                            tool_call_delta: None,
-                            is_finished: true,
-                            finish_reason: Some(format!(
-                                "stream_error:prompt_blocked:{}",
-                                block_reason.to_ascii_lowercase()
-                            )),
-                            provider_response_id: None,
-                            provider_response_status: None,
-                        })
-                        .await;
-                    emitted_final = true;
-                    break;
-                }
-
-                selected_candidate_index = select_stream_candidate_index(
-                    selected_candidate_index,
-                    emitted_payload,
-                    &gemini_chunk.candidates,
-                );
-
-                for (candidate_position, candidate) in
-                    gemini_chunk.candidates.into_iter().enumerate()
-                {
-                    if !should_consume_stream_candidate(
-                        selected_candidate_index,
-                        candidate_position,
-                        candidate.index,
-                    ) {
-                        continue;
-                    }
-                    let finish_reason = candidate.finish_reason.clone();
-                    if let Some(content) = candidate.content {
-                        for part in content.parts {
-                            if let Some(text) = part.text {
-                                emitted_payload = true;
-                                let stream_chunk = UnifiedStreamChunk {
-                                    text: Some(text),
-                                    thinking: None,
-                                    thinking_signature: None,
-                                    redacted_thinking: None,
-                                    usage: None,
-                                    sequence_number: None,
-                                    tool_call_delta: None,
-                                    is_finished: false,
-                                    finish_reason: None,
-                                    provider_response_id: None,
-                                    provider_response_status: None,
-                                };
-                                if convert_tx.send(stream_chunk).await.is_err() {
-                                    return;
-                                }
-                            }
-
-                            if let Some(fc) = part.function_call {
-                                let tool_call_index = next_tool_call_index;
-                                next_tool_call_index = next_tool_call_index.saturating_add(1);
-                                emitted_payload = true;
-                                let stream_chunk = UnifiedStreamChunk {
-                                    text: None,
-                                    thinking: None,
-                                    thinking_signature: None,
-                                    redacted_thinking: None,
-                                    usage: None,
-                                    sequence_number: None,
-                                    tool_call_delta: Some(crate::ToolCallDelta {
-                                        index: tool_call_index,
-                                        id: None,
-                                        name: Some(fc.name),
-                                        arguments_delta: Some(fc.args.to_string()),
-                                        arguments: None,
-                                    }),
-                                    is_finished: false,
-                                    finish_reason: None,
-                                    provider_response_id: None,
-                                    provider_response_status: None,
-                                };
-                                if convert_tx.send(stream_chunk).await.is_err() {
-                                    return;
-                                }
-                            }
-                        }
-                    }
-
-                    if finish_reason.is_some() {
-                        let final_chunk = UnifiedStreamChunk {
-                            text: None,
-                            thinking: None,
-                            thinking_signature: None,
-                            redacted_thinking: None,
-                            usage: latest_usage,
-                            sequence_number: None,
-                            tool_call_delta: None,
-                            is_finished: true,
-                            finish_reason: finish_reason.map(normalize_stream_finish_reason),
-                            provider_response_id: None,
-                            provider_response_status: None,
-                        };
-                        emitted_final = true;
-                        let _ = convert_tx.send(final_chunk).await;
-                    }
-                }
-            }
-
-            // If stream ended without finish event:
-            // - emit final only when we already saw payload, so runtime can finalize streamed output.
-            // - emit nothing on totally empty stream, so runtime can fallback to non-streaming generate().
-            if !emitted_final && emitted_payload {
-                let _ = convert_tx
-                    .send(UnifiedStreamChunk {
-                        text: None,
-                        thinking: None,
-                        thinking_signature: None,
-                        redacted_thinking: None,
-                        usage: latest_usage,
-                        sequence_number: None,
-                        tool_call_delta: None,
-                        is_finished: true,
-                        finish_reason: Some("stream_closed".to_string()),
-                        provider_response_id: None,
-                        provider_response_status: None,
-                    })
-                    .await;
-            }
-        });
+        tokio::spawn(streaming::project_chunks(gemini_rx, tx));
 
         let mut stream_client = GoogleGeminiGenerateContentClient::with_params(
             &self.project_id,
@@ -1043,331 +775,4 @@ impl LlmProvider for GoogleGeminiGenerateContentClient {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_gemini_client_with_params() {
-        let client = GoogleGeminiGenerateContentClient::with_params(
-            "test-project",
-            "us-central1",
-            "gemini-2.0-flash",
-        );
-        // Just verify it compiles and creates
-        drop(client);
-    }
-
-    #[test]
-    fn test_part_text() {
-        let part = Part::text("Hello, world!");
-        assert_eq!(part.text, Some("Hello, world!".to_string()));
-        assert!(part.function_call.is_none());
-        assert!(part.function_response.is_none());
-    }
-
-    #[test]
-    fn test_part_function_call() {
-        let args = serde_json::json!({"query": "test"});
-        let part = Part::function_call("web_search", args.clone());
-        assert!(part.text.is_none());
-        assert!(part.function_call.is_some());
-        let fc = part.function_call.unwrap();
-        assert_eq!(fc.name, "web_search");
-        assert_eq!(fc.args, args);
-    }
-
-    #[test]
-    fn test_part_function_response() {
-        let response = serde_json::json!({"result": "success"});
-        let part = Part::function_response("my_tool", response.clone());
-        assert!(part.text.is_none());
-        assert!(part.function_response.is_some());
-        let fr = part.function_response.unwrap();
-        assert_eq!(fr.name, "my_tool");
-        assert_eq!(fr.response, response);
-    }
-
-    #[test]
-    fn test_content_user() {
-        let parts = vec![Part::text("Hello")];
-        let content = Content::user(parts);
-        assert_eq!(content.role, Some("user".to_string()));
-        assert_eq!(content.parts.len(), 1);
-    }
-
-    #[test]
-    fn test_content_model() {
-        let parts = vec![Part::text("Response")];
-        let content = Content::model(parts);
-        assert_eq!(content.role, Some("model".to_string()));
-        assert_eq!(content.parts.len(), 1);
-    }
-
-    #[test]
-    fn test_content_function() {
-        let parts = vec![Part::function_response("tool", serde_json::json!({}))];
-        let content = Content::function(parts);
-        assert_eq!(content.role, Some("function".to_string()));
-        assert_eq!(content.parts.len(), 1);
-    }
-
-    #[test]
-    fn test_generation_config_default() {
-        let config = GenerationConfig::default();
-        assert!(config.temperature.is_none());
-        assert!(config.max_output_tokens.is_none());
-        assert!(config.top_p.is_none());
-        assert!(config.top_k.is_none());
-        assert!(config.thinking_config.is_none());
-    }
-
-    #[test]
-    fn test_build_gemini_thinking_config_maps_gemini_3_effort_to_level() {
-        let config =
-            build_gemini_thinking_config("gemini-3-flash-preview", Some(ReasoningEffort::Medium))
-                .unwrap()
-                .unwrap();
-
-        assert_eq!(config.thinking_level.as_deref(), Some("medium"));
-        assert_eq!(config.thinking_budget, None);
-    }
-
-    #[test]
-    fn test_build_gemini_thinking_config_maps_gemini_25_effort_to_budget() {
-        let config = build_gemini_thinking_config("gemini-2.5-flash", Some(ReasoningEffort::High))
-            .unwrap()
-            .unwrap();
-
-        assert_eq!(config.thinking_level, None);
-        assert_eq!(config.thinking_budget, Some(8192));
-    }
-
-    #[test]
-    fn test_build_gemini_thinking_config_omits_budget_when_effort_unset() {
-        let config = build_gemini_thinking_config("gemini-2.5-flash", None).unwrap();
-
-        assert!(config.is_none());
-    }
-
-    #[test]
-    fn test_select_stream_candidate_index_prefers_zero_then_falls_back() {
-        let non_zero_candidates = vec![
-            StreamCandidate {
-                content: None,
-                finish_reason: None,
-                index: Some(3),
-            },
-            StreamCandidate {
-                content: None,
-                finish_reason: None,
-                index: Some(4),
-            },
-        ];
-        assert_eq!(
-            select_stream_candidate_index(None, false, &non_zero_candidates),
-            Some(3)
-        );
-
-        let with_zero = vec![
-            StreamCandidate {
-                content: None,
-                finish_reason: None,
-                index: Some(2),
-            },
-            StreamCandidate {
-                content: None,
-                finish_reason: None,
-                index: Some(0),
-            },
-        ];
-        assert_eq!(
-            select_stream_candidate_index(None, false, &with_zero),
-            Some(0)
-        );
-        assert_eq!(
-            select_stream_candidate_index(Some(3), false, &with_zero),
-            Some(0)
-        );
-        assert_eq!(
-            select_stream_candidate_index(Some(3), true, &with_zero),
-            Some(3)
-        );
-    }
-
-    #[test]
-    fn test_should_consume_stream_candidate_uses_selected_index_or_position() {
-        assert!(should_consume_stream_candidate(Some(0), 1, Some(0)));
-        assert!(!should_consume_stream_candidate(Some(0), 0, Some(1)));
-        assert!(should_consume_stream_candidate(None, 0, None));
-        assert!(!should_consume_stream_candidate(None, 1, None));
-    }
-
-    #[test]
-    fn test_normalize_stream_finish_reason_maps_blocking_reasons_to_stream_error() {
-        assert_eq!(
-            normalize_stream_finish_reason("SAFETY".to_string()),
-            "stream_error:safety"
-        );
-        assert_eq!(
-            normalize_stream_finish_reason("RECITATION".to_string()),
-            "stream_error:recitation"
-        );
-        assert_eq!(normalize_stream_finish_reason("STOP".to_string()), "STOP");
-    }
-
-    #[test]
-    fn test_select_primary_candidate_prefers_index_zero() {
-        let candidates = vec![
-            Candidate {
-                content: Some(Content::model(vec![Part::text("secondary")])),
-                finish_reason: Some("STOP".to_string()),
-                index: Some(2),
-                safety_ratings: vec![],
-            },
-            Candidate {
-                content: Some(Content::model(vec![Part::text("primary")])),
-                finish_reason: Some("STOP".to_string()),
-                index: Some(0),
-                safety_ratings: vec![],
-            },
-        ];
-
-        let selected = select_primary_candidate(&candidates).expect("expected candidate");
-        assert_eq!(selected.index, Some(0));
-        assert_eq!(
-            selected
-                .content
-                .as_ref()
-                .and_then(|c| c.parts.first())
-                .and_then(|p| p.text.as_deref()),
-            Some("primary")
-        );
-    }
-
-    #[test]
-    fn test_generate_content_request_serialization() {
-        let request = GoogleGeminiGenerateContentRequest {
-            contents: vec![Content::user(vec![Part::text("Hello")])],
-            system_instruction: None,
-            tools: None,
-            generation_config: Some(GenerationConfig {
-                temperature: Some(0.7),
-                max_output_tokens: Some(100),
-                top_p: None,
-                top_k: None,
-                thinking_config: Some(ThinkingConfig {
-                    thinking_level: Some("low".to_string()),
-                    thinking_budget: None,
-                }),
-            }),
-        };
-
-        let json = serde_json::to_string(&request).unwrap();
-        assert!(json.contains("contents"));
-        assert!(json.contains("generationConfig"));
-        assert!(json.contains("thinkingConfig"));
-        assert!(json.contains("thinkingLevel"));
-        assert!(json.contains("0.7"));
-    }
-
-    #[test]
-    fn test_generate_content_response_deserialization() {
-        let json = r#"{
-            "candidates": [{
-                "content": {
-                    "role": "model",
-                    "parts": [{"text": "Hello!"}]
-                },
-                "finishReason": "STOP"
-            }],
-            "usageMetadata": {
-                "promptTokenCount": 10,
-                "candidatesTokenCount": 5,
-                "totalTokenCount": 15
-            }
-        }"#;
-
-        let response: GoogleGeminiGenerateContentResponse = serde_json::from_str(json).unwrap();
-        assert_eq!(response.candidates.len(), 1);
-        assert!(response.usage_metadata.is_some());
-        let usage = response.usage_metadata.unwrap();
-        assert_eq!(usage.prompt_token_count, Some(10));
-        assert_eq!(usage.total_token_count, Some(15));
-    }
-
-    #[test]
-    fn test_safety_rating() {
-        let json = r#"{
-            "category": "HARM_CATEGORY_SEXUALLY_EXPLICIT",
-            "probability": "NEGLIGIBLE"
-        }"#;
-
-        let rating: SafetyRating = serde_json::from_str(json).unwrap();
-        assert_eq!(rating.category, "HARM_CATEGORY_SEXUALLY_EXPLICIT");
-        assert_eq!(rating.probability, "NEGLIGIBLE");
-    }
-
-    #[test]
-    fn test_prompt_feedback_deserialization() {
-        let json = r#"{
-            "blockReason": "SAFETY",
-            "safetyRatings": [
-                {"category": "HARM_CATEGORY_HARASSMENT", "probability": "HIGH"}
-            ]
-        }"#;
-
-        let feedback: PromptFeedback = serde_json::from_str(json).unwrap();
-        assert_eq!(feedback.block_reason, Some("SAFETY".to_string()));
-        assert_eq!(feedback.safety_ratings.len(), 1);
-    }
-
-    #[test]
-    fn test_function_declaration() {
-        let decl = FunctionDeclaration {
-            name: "test_function".to_string(),
-            description: "A test function".to_string(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "arg1": {"type": "string"}
-                }
-            }),
-        };
-
-        assert_eq!(decl.name, "test_function");
-        assert_eq!(decl.description, "A test function");
-    }
-
-    #[test]
-    fn test_tool_serialization() {
-        let tool = Tool {
-            function_declarations: vec![FunctionDeclaration {
-                name: "func1".to_string(),
-                description: "Function 1".to_string(),
-                parameters: serde_json::json!({}),
-            }],
-        };
-
-        let json = serde_json::to_string(&tool).unwrap();
-        assert!(json.contains("functionDeclarations"));
-    }
-
-    #[test]
-    fn test_stream_chunk_deserialization() {
-        let json = r#"{
-            "candidates": [{
-                "content": {
-                    "role": "model",
-                    "parts": [{"text": "Hello"}]
-                },
-                "finishReason": null,
-                "index": 0
-            }],
-            "usageMetadata": null
-        }"#;
-
-        let chunk: StreamChunk = serde_json::from_str(json).unwrap();
-        assert_eq!(chunk.candidates.len(), 1);
-        assert!(chunk.usage_metadata.is_none());
-    }
-}
+mod tests;
